Rework *when* compile classpath and task happens
The 'javadoc' task reads the source set's compileClasspath before we
*replace* it our compileOnly configuration. By configuring the compile
classpath when the JavaBasePlugin is applied, it can replace the compile
classpath before the javadoc task is configured by the JavaPlugin.

Because the compile tasks are created and configured in the JavaBasePlugin
too (LegacyJavaComponentPlugin actually, installed by JavaBasePlugin),
we move the task's processorpath and generated source output configuration
there.

This change will also configure all source sets and compilation tasks, not
just the "main" and "test" source sets; in case another plugin adds a new
one.