Modify .idea/compiler.xml if present
While `gradle idea` creates a "file-based" IntelliJ project, with a *.ipr file containing the project-wide annotation-processing configuration that we need to set, the integrated gradle support in IntelliJ defaults to the more recent (though still over 6 years old) directory-based configuration, storing the relevant settings in `.idea/compiler.xml`. Supporting the latter is as simple as modifying that file, if present, allowing users to just open a project's `gradle.build` file in IntelliJ and start work.