Make everything configurable through Gradle "conventions"
This is based on https://github.com/gradle/gradle/blob/master/design-docs/java-annotation-processing.md
except that proposed new CompileOptions properties live on an
'aptOptions' property, because CompileOptions is not "convention
aware".

Note that this is a breaking change as -processorpath is now only
set if the path would be non-empty.