Refactor AptPluginSpec to only use public APIs
Unit tests run in the current version of Gradle, so take
advantage of this to test that the native properties are
correctly configured by the plugin. This also removes a
few tests that were specific to previous Gradle versions,
and were now actually testing Gradle's own behavior
rather than the plugin.
As a side effect, this no longer tests some expected
behaviors in previous Gradle versions; but these should
be tested in integration tests.
This is in preparation for a overhaul of the plugin
implementation.