Fix eclipse integration test with Gradle 4.6
The cleanEclipse task in Gradle 4.6 is always up-to-date, as
it actually does nothing (but depends on other cleaning tasks;
this was already the case before, but the task then had a
successful outcome). Because we don't actually really care
about that task's outcome, we can just stop checking its
outcome (as well as the one for the eclipse task).
Similarly, stop checking the outcome of 'idea' tasks.