Fix potential regression in apt-eclipse in v0.20
introduced when migrating to using lazy properties internally:
the eclipse.jdt.apt.processorOptions is a Map and modifying it
possibly modified the aptOptions.processorArgs that was used
as the default value (even though I haven't reproduced the
potential issue).

Anyway, this will get rid of the warning that was emitted when
previously using a Property<Map> in Gradle 5.1.

Ideally, we'd move to using a MapProperty but this requires
Gradle 5.1 and would break the API.

Fixes #94